+++ packages/camera/camera_android_camerax/test/android_camera_camerax_test.mocks.dart
@@ -3,35 +3,39 @@
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i11;
-import 'dart:typed_data' as _i20;
+import 'dart:async' as _i14;
+import 'dart:typed_data' as _i23;
 
-import 'package:camera_android_camerax/src/analyzer.dart' as _i16;
+import 'package:camera_android_camerax/src/analyzer.dart' as _i19;
 import 'package:camera_android_camerax/src/camera.dart' as _i7;
 import 'package:camera_android_camerax/src/camera_info.dart' as _i2;
-import 'package:camera_android_camerax/src/camera_selector.dart' as _i14;
-import 'package:camera_android_camerax/src/camera_state.dart' as _i12;
+import 'package:camera_android_camerax/src/camera_selector.dart' as _i17;
+import 'package:camera_android_camerax/src/camera_state.dart' as _i15;
 import 'package:camera_android_camerax/src/camerax_library.g.dart' as _i6;
 import 'package:camera_android_camerax/src/exposure_state.dart' as _i4;
-import 'package:camera_android_camerax/src/image_analysis.dart' as _i15;
-import 'package:camera_android_camerax/src/image_capture.dart' as _i17;
-import 'package:camera_android_camerax/src/image_proxy.dart' as _i18;
+import 'package:camera_android_camerax/src/image_analysis.dart' as _i18;
+import 'package:camera_android_camerax/src/image_capture.dart' as _i20;
+import 'package:camera_android_camerax/src/image_proxy.dart' as _i21;
 import 'package:camera_android_camerax/src/live_data.dart' as _i3;
-import 'package:camera_android_camerax/src/observer.dart' as _i25;
-import 'package:camera_android_camerax/src/plane_proxy.dart' as _i19;
-import 'package:camera_android_camerax/src/preview.dart' as _i21;
+import 'package:camera_android_camerax/src/observer.dart' as _i29;
+import 'package:camera_android_camerax/src/pending_recording.dart' as _i8;
+import 'package:camera_android_camerax/src/plane_proxy.dart' as _i22;
+import 'package:camera_android_camerax/src/preview.dart' as _i24;
 import 'package:camera_android_camerax/src/process_camera_provider.dart'
-    as _i22;
-import 'package:camera_android_camerax/src/use_case.dart' as _i23;
-import 'package:camera_android_camerax/src/zoom_state.dart' as _i13;
+    as _i25;
+import 'package:camera_android_camerax/src/recorder.dart' as _i10;
+import 'package:camera_android_camerax/src/recording.dart' as _i9;
+import 'package:camera_android_camerax/src/use_case.dart' as _i26;
+import 'package:camera_android_camerax/src/video_capture.dart' as _i27;
+import 'package:camera_android_camerax/src/zoom_state.dart' as _i16;
 import 'package:camera_platform_interface/camera_platform_interface.dart'
     as _i5;
-import 'package:flutter/foundation.dart' as _i10;
-import 'package:flutter/services.dart' as _i9;
-import 'package:flutter/widgets.dart' as _i8;
+import 'package:flutter/foundation.dart' as _i13;
+import 'package:flutter/services.dart' as _i12;
+import 'package:flutter/widgets.dart' as _i11;
 import 'package:mockito/mockito.dart' as _i1;
 
-import 'test_camerax_library.g.dart' as _i24;
+import 'test_camerax_library.g.dart' as _i28;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -118,8 +122,39 @@ class _FakeCamera_6 extends _i1.SmartFake implements _i7.Camera {
         );
 }
 
-class _FakeWidget_7 extends _i1.SmartFake implements _i8.Widget {
-  _FakeWidget_7(
+class _FakePendingRecording_7 extends _i1.SmartFake
+    implements _i8.PendingRecording {
+  _FakePendingRecording_7(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeRecording_8 extends _i1.SmartFake implements _i9.Recording {
+  _FakeRecording_8(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeRecorder_9 extends _i1.SmartFake implements _i10.Recorder {
+  _FakeRecorder_9(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeWidget_10 extends _i1.SmartFake implements _i11.Widget {
+  _FakeWidget_10(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -128,13 +163,14 @@ class _FakeWidget_7 extends _i1.SmartFake implements _i8.Widget {
         );
 
   @override
-  String toString({_i9.DiagnosticLevel? minLevel = _i9.DiagnosticLevel.info}) =>
+  String toString(
+          {_i12.DiagnosticLevel? minLevel = _i12.DiagnosticLevel.info}) =>
       super.toString();
 }
 
-class _FakeInheritedWidget_8 extends _i1.SmartFake
-    implements _i8.InheritedWidget {
-  _FakeInheritedWidget_8(
+class _FakeInheritedWidget_11 extends _i1.SmartFake
+    implements _i11.InheritedWidget {
+  _FakeInheritedWidget_11(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -143,13 +179,14 @@ class _FakeInheritedWidget_8 extends _i1.SmartFake
         );
 
   @override
-  String toString({_i9.DiagnosticLevel? minLevel = _i9.DiagnosticLevel.info}) =>
+  String toString(
+          {_i12.DiagnosticLevel? minLevel = _i12.DiagnosticLevel.info}) =>
       super.toString();
 }
 
-class _FakeDiagnosticsNode_9 extends _i1.SmartFake
-    implements _i10.DiagnosticsNode {
-  _FakeDiagnosticsNode_9(
+class _FakeDiagnosticsNode_12 extends _i1.SmartFake
+    implements _i13.DiagnosticsNode {
+  _FakeDiagnosticsNode_12(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -159,8 +196,8 @@ class _FakeDiagnosticsNode_9 extends _i1.SmartFake
 
   @override
   String toString({
-    _i10.TextTreeConfiguration? parentConfiguration,
-    _i9.DiagnosticLevel? minLevel = _i9.DiagnosticLevel.info,
+    _i13.TextTreeConfiguration? parentConfiguration,
+    _i12.DiagnosticLevel? minLevel = _i12.DiagnosticLevel.info,
   }) =>
       super.toString();
 }
@@ -170,12 +207,12 @@ class _FakeDiagnosticsNode_9 extends _i1.SmartFake
 /// See the documentation for Mockito's code generation for more information.
 class MockCamera extends _i1.Mock implements _i7.Camera {
   @override
-  _i11.Future<_i2.CameraInfo> getCameraInfo() => (super.noSuchMethod(
+  _i14.Future<_i2.CameraInfo> getCameraInfo() => (super.noSuchMethod(
         Invocation.method(
           #getCameraInfo,
           [],
         ),
-        returnValue: _i11.Future<_i2.CameraInfo>.value(_FakeCameraInfo_0(
+        returnValue: _i14.Future<_i2.CameraInfo>.value(_FakeCameraInfo_0(
           this,
           Invocation.method(
             #getCameraInfo,
@@ -183,14 +220,14 @@ class MockCamera extends _i1.Mock implements _i7.Camera {
           ),
         )),
         returnValueForMissingStub:
-            _i11.Future<_i2.CameraInfo>.value(_FakeCameraInfo_0(
+            _i14.Future<_i2.CameraInfo>.value(_FakeCameraInfo_0(
           this,
           Invocation.method(
             #getCameraInfo,
             [],
           ),
         )),
-      ) as _i11.Future<_i2.CameraInfo>);
+      ) as _i14.Future<_i2.CameraInfo>);
 }
 
 /// A class which mocks [CameraInfo].
@@ -198,23 +235,23 @@ class MockCamera extends _i1.Mock implements _i7.Camera {
 /// See the documentation for Mockito's code generation for more information.
 class MockCameraInfo extends _i1.Mock implements _i2.CameraInfo {
   @override
-  _i11.Future<int> getSensorRotationDegrees() => (super.noSuchMethod(
+  _i14.Future<int> getSensorRotationDegrees() => (super.noSuchMethod(
         Invocation.method(
           #getSensorRotationDegrees,
           [],
         ),
-        returnValue: _i11.Future<int>.value(0),
-        returnValueForMissingStub: _i11.Future<int>.value(0),
-      ) as _i11.Future<int>);
+        returnValue: _i14.Future<int>.value(0),
+        returnValueForMissingStub: _i14.Future<int>.value(0),
+      ) as _i14.Future<int>);
   @override
-  _i11.Future<_i3.LiveData<_i12.CameraState>> getCameraState() =>
+  _i14.Future<_i3.LiveData<_i15.CameraState>> getCameraState() =>
       (super.noSuchMethod(
         Invocation.method(
           #getCameraState,
           [],
         ),
-        returnValue: _i11.Future<_i3.LiveData<_i12.CameraState>>.value(
-            _FakeLiveData_1<_i12.CameraState>(
+        returnValue: _i14.Future<_i3.LiveData<_i15.CameraState>>.value(
+            _FakeLiveData_1<_i15.CameraState>(
           this,
           Invocation.method(
             #getCameraState,
@@ -222,22 +259,22 @@ class MockCameraInfo extends _i1.Mock implements _i2.CameraInfo {
           ),
         )),
         returnValueForMissingStub:
-            _i11.Future<_i3.LiveData<_i12.CameraState>>.value(
-                _FakeLiveData_1<_i12.CameraState>(
+            _i14.Future<_i3.LiveData<_i15.CameraState>>.value(
+                _FakeLiveData_1<_i15.CameraState>(
           this,
           Invocation.method(
             #getCameraState,
             [],
           ),
         )),
-      ) as _i11.Future<_i3.LiveData<_i12.CameraState>>);
+      ) as _i14.Future<_i3.LiveData<_i15.CameraState>>);
   @override
-  _i11.Future<_i4.ExposureState> getExposureState() => (super.noSuchMethod(
+  _i14.Future<_i4.ExposureState> getExposureState() => (super.noSuchMethod(
         Invocation.method(
           #getExposureState,
           [],
         ),
-        returnValue: _i11.Future<_i4.ExposureState>.value(_FakeExposureState_2(
+        returnValue: _i14.Future<_i4.ExposureState>.value(_FakeExposureState_2(
           this,
           Invocation.method(
             #getExposureState,
@@ -245,23 +282,23 @@ class MockCameraInfo extends _i1.Mock implements _i2.CameraInfo {
           ),
         )),
         returnValueForMissingStub:
-            _i11.Future<_i4.ExposureState>.value(_FakeExposureState_2(
+            _i14.Future<_i4.ExposureState>.value(_FakeExposureState_2(
           this,
           Invocation.method(
             #getExposureState,
             [],
           ),
         )),
-      ) as _i11.Future<_i4.ExposureState>);
+      ) as _i14.Future<_i4.ExposureState>);
   @override
-  _i11.Future<_i3.LiveData<_i13.ZoomState>> getZoomState() =>
+  _i14.Future<_i3.LiveData<_i16.ZoomState>> getZoomState() =>
       (super.noSuchMethod(
         Invocation.method(
           #getZoomState,
           [],
         ),
-        returnValue: _i11.Future<_i3.LiveData<_i13.ZoomState>>.value(
-            _FakeLiveData_1<_i13.ZoomState>(
+        returnValue: _i14.Future<_i3.LiveData<_i16.ZoomState>>.value(
+            _FakeLiveData_1<_i16.ZoomState>(
           this,
           Invocation.method(
             #getZoomState,
@@ -269,15 +306,15 @@ class MockCameraInfo extends _i1.Mock implements _i2.CameraInfo {
           ),
         )),
         returnValueForMissingStub:
-            _i11.Future<_i3.LiveData<_i13.ZoomState>>.value(
-                _FakeLiveData_1<_i13.ZoomState>(
+            _i14.Future<_i3.LiveData<_i16.ZoomState>>.value(
+                _FakeLiveData_1<_i16.ZoomState>(
           this,
           Invocation.method(
             #getZoomState,
             [],
           ),
         )),
-      ) as _i11.Future<_i3.LiveData<_i13.ZoomState>>);
+      ) as _i14.Future<_i3.LiveData<_i16.ZoomState>>);
 }
 
 /// A class which mocks [CameraImageData].
@@ -320,19 +357,19 @@ class MockCameraImageData extends _i1.Mock implements _i5.CameraImageData {
 /// A class which mocks [CameraSelector].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockCameraSelector extends _i1.Mock implements _i14.CameraSelector {
+class MockCameraSelector extends _i1.Mock implements _i17.CameraSelector {
   @override
-  _i11.Future<List<_i2.CameraInfo>> filter(List<_i2.CameraInfo>? cameraInfos) =>
+  _i14.Future<List<_i2.CameraInfo>> filter(List<_i2.CameraInfo>? cameraInfos) =>
       (super.noSuchMethod(
         Invocation.method(
           #filter,
           [cameraInfos],
         ),
         returnValue:
-            _i11.Future<List<_i2.CameraInfo>>.value(<_i2.CameraInfo>[]),
+            _i14.Future<List<_i2.CameraInfo>>.value(<_i2.CameraInfo>[]),
         returnValueForMissingStub:
-            _i11.Future<List<_i2.CameraInfo>>.value(<_i2.CameraInfo>[]),
-      ) as _i11.Future<List<_i2.CameraInfo>>);
+            _i14.Future<List<_i2.CameraInfo>>.value(<_i2.CameraInfo>[]),
+      ) as _i14.Future<List<_i2.CameraInfo>>);
 }
 
 /// A class which mocks [ExposureState].
@@ -363,55 +400,55 @@ class MockExposureState extends _i1.Mock implements _i4.ExposureState {
 /// A class which mocks [ImageAnalysis].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockImageAnalysis extends _i1.Mock implements _i15.ImageAnalysis {
+class MockImageAnalysis extends _i1.Mock implements _i18.ImageAnalysis {
   @override
-  _i11.Future<void> setAnalyzer(_i16.Analyzer? analyzer) => (super.noSuchMethod(
+  _i14.Future<void> setAnalyzer(_i19.Analyzer? analyzer) => (super.noSuchMethod(
         Invocation.method(
           #setAnalyzer,
           [analyzer],
         ),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
   @override
-  _i11.Future<void> clearAnalyzer() => (super.noSuchMethod(
+  _i14.Future<void> clearAnalyzer() => (super.noSuchMethod(
         Invocation.method(
           #clearAnalyzer,
           [],
         ),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
 }
 
 /// A class which mocks [ImageCapture].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockImageCapture extends _i1.Mock implements _i17.ImageCapture {
+class MockImageCapture extends _i1.Mock implements _i20.ImageCapture {
   @override
-  _i11.Future<void> setFlashMode(int? newFlashMode) => (super.noSuchMethod(
+  _i14.Future<void> setFlashMode(int? newFlashMode) => (super.noSuchMethod(
         Invocation.method(
           #setFlashMode,
           [newFlashMode],
         ),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
   @override
-  _i11.Future<String> takePicture() => (super.noSuchMethod(
+  _i14.Future<String> takePicture() => (super.noSuchMethod(
         Invocation.method(
           #takePicture,
           [],
         ),
-        returnValue: _i11.Future<String>.value(''),
-        returnValueForMissingStub: _i11.Future<String>.value(''),
-      ) as _i11.Future<String>);
+        returnValue: _i14.Future<String>.value(''),
+        returnValueForMissingStub: _i14.Future<String>.value(''),
+      ) as _i14.Future<String>);
 }
 
 /// A class which mocks [ImageProxy].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockImageProxy extends _i1.Mock implements _i18.ImageProxy {
+class MockImageProxy extends _i1.Mock implements _i21.ImageProxy {
   @override
   int get format => (super.noSuchMethod(
         Invocation.getter(#format),
@@ -431,37 +468,37 @@ class MockImageProxy extends _i1.Mock implements _i18.ImageProxy {
         returnValueForMissingStub: 0,
       ) as int);
   @override
-  _i11.Future<List<_i19.PlaneProxy>> getPlanes() => (super.noSuchMethod(
+  _i14.Future<List<_i22.PlaneProxy>> getPlanes() => (super.noSuchMethod(
         Invocation.method(
           #getPlanes,
           [],
         ),
         returnValue:
-            _i11.Future<List<_i19.PlaneProxy>>.value(<_i19.PlaneProxy>[]),
+            _i14.Future<List<_i22.PlaneProxy>>.value(<_i22.PlaneProxy>[]),
         returnValueForMissingStub:
-            _i11.Future<List<_i19.PlaneProxy>>.value(<_i19.PlaneProxy>[]),
-      ) as _i11.Future<List<_i19.PlaneProxy>>);
+            _i14.Future<List<_i22.PlaneProxy>>.value(<_i22.PlaneProxy>[]),
+      ) as _i14.Future<List<_i22.PlaneProxy>>);
   @override
-  _i11.Future<void> close() => (super.noSuchMethod(
+  _i14.Future<void> close() => (super.noSuchMethod(
         Invocation.method(
           #close,
           [],
         ),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
 }
 
 /// A class which mocks [PlaneProxy].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockPlaneProxy extends _i1.Mock implements _i19.PlaneProxy {
+class MockPlaneProxy extends _i1.Mock implements _i22.PlaneProxy {
   @override
-  _i20.Uint8List get buffer => (super.noSuchMethod(
+  _i23.Uint8List get buffer => (super.noSuchMethod(
         Invocation.getter(#buffer),
-        returnValue: _i20.Uint8List(0),
-        returnValueForMissingStub: _i20.Uint8List(0),
-      ) as _i20.Uint8List);
+        returnValue: _i23.Uint8List(0),
+        returnValueForMissingStub: _i23.Uint8List(0),
+      ) as _i23.Uint8List);
   @override
   int get pixelStride => (super.noSuchMethod(
         Invocation.getter(#pixelStride),
@@ -479,16 +516,16 @@ class MockPlaneProxy extends _i1.Mock implements _i19.PlaneProxy {
 /// A class which mocks [Preview].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockPreview extends _i1.Mock implements _i21.Preview {
+class MockPreview extends _i1.Mock implements _i24.Preview {
   @override
-  _i11.Future<int> setSurfaceProvider() => (super.noSuchMethod(
+  _i14.Future<int> setSurfaceProvider() => (super.noSuchMethod(
         Invocation.method(
           #setSurfaceProvider,
           [],
         ),
-        returnValue: _i11.Future<int>.value(0),
-        returnValueForMissingStub: _i11.Future<int>.value(0),
-      ) as _i11.Future<int>);
+        returnValue: _i14.Future<int>.value(0),
+        returnValueForMissingStub: _i14.Future<int>.value(0),
+      ) as _i14.Future<int>);
   @override
   void releaseFlutterSurfaceTexture() => super.noSuchMethod(
         Invocation.method(
@@ -498,13 +535,13 @@ class MockPreview extends _i1.Mock implements _i21.Preview {
         returnValueForMissingStub: null,
       );
   @override
-  _i11.Future<_i6.ResolutionInfo> getResolutionInfo() => (super.noSuchMethod(
+  _i14.Future<_i6.ResolutionInfo> getResolutionInfo() => (super.noSuchMethod(
         Invocation.method(
           #getResolutionInfo,
           [],
         ),
         returnValue:
-            _i11.Future<_i6.ResolutionInfo>.value(_FakeResolutionInfo_5(
+            _i14.Future<_i6.ResolutionInfo>.value(_FakeResolutionInfo_5(
           this,
           Invocation.method(
             #getResolutionInfo,
@@ -512,37 +549,37 @@ class MockPreview extends _i1.Mock implements _i21.Preview {
           ),
         )),
         returnValueForMissingStub:
-            _i11.Future<_i6.ResolutionInfo>.value(_FakeResolutionInfo_5(
+            _i14.Future<_i6.ResolutionInfo>.value(_FakeResolutionInfo_5(
           this,
           Invocation.method(
             #getResolutionInfo,
             [],
           ),
         )),
-      ) as _i11.Future<_i6.ResolutionInfo>);
+      ) as _i14.Future<_i6.ResolutionInfo>);
 }
 
 /// A class which mocks [ProcessCameraProvider].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockProcessCameraProvider extends _i1.Mock
-    implements _i22.ProcessCameraProvider {
+    implements _i25.ProcessCameraProvider {
   @override
-  _i11.Future<List<_i2.CameraInfo>> getAvailableCameraInfos() =>
+  _i14.Future<List<_i2.CameraInfo>> getAvailableCameraInfos() =>
       (super.noSuchMethod(
         Invocation.method(
           #getAvailableCameraInfos,
           [],
         ),
         returnValue:
-            _i11.Future<List<_i2.CameraInfo>>.value(<_i2.CameraInfo>[]),
+            _i14.Future<List<_i2.CameraInfo>>.value(<_i2.CameraInfo>[]),
         returnValueForMissingStub:
-            _i11.Future<List<_i2.CameraInfo>>.value(<_i2.CameraInfo>[]),
-      ) as _i11.Future<List<_i2.CameraInfo>>);
+            _i14.Future<List<_i2.CameraInfo>>.value(<_i2.CameraInfo>[]),
+      ) as _i14.Future<List<_i2.CameraInfo>>);
   @override
-  _i11.Future<_i7.Camera> bindToLifecycle(
-    _i14.CameraSelector? cameraSelector,
-    List<_i23.UseCase>? useCases,
+  _i14.Future<_i7.Camera> bindToLifecycle(
+    _i17.CameraSelector? cameraSelector,
+    List<_i26.UseCase>? useCases,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -552,7 +589,7 @@ class MockProcessCameraProvider extends _i1.Mock
             useCases,
           ],
         ),
-        returnValue: _i11.Future<_i7.Camera>.value(_FakeCamera_6(
+        returnValue: _i14.Future<_i7.Camera>.value(_FakeCamera_6(
           this,
           Invocation.method(
             #bindToLifecycle,
@@ -562,7 +599,7 @@ class MockProcessCameraProvider extends _i1.Mock
             ],
           ),
         )),
-        returnValueForMissingStub: _i11.Future<_i7.Camera>.value(_FakeCamera_6(
+        returnValueForMissingStub: _i14.Future<_i7.Camera>.value(_FakeCamera_6(
           this,
           Invocation.method(
             #bindToLifecycle,
@@ -572,18 +609,18 @@ class MockProcessCameraProvider extends _i1.Mock
             ],
           ),
         )),
-      ) as _i11.Future<_i7.Camera>);
+      ) as _i14.Future<_i7.Camera>);
   @override
-  _i11.Future<bool> isBound(_i23.UseCase? useCase) => (super.noSuchMethod(
+  _i14.Future<bool> isBound(_i26.UseCase? useCase) => (super.noSuchMethod(
         Invocation.method(
           #isBound,
           [useCase],
         ),
-        returnValue: _i11.Future<bool>.value(false),
-        returnValueForMissingStub: _i11.Future<bool>.value(false),
-      ) as _i11.Future<bool>);
+        returnValue: _i14.Future<bool>.value(false),
+        returnValueForMissingStub: _i14.Future<bool>.value(false),
+      ) as _i14.Future<bool>);
   @override
-  void unbind(List<_i23.UseCase>? useCases) => super.noSuchMethod(
+  void unbind(List<_i26.UseCase>? useCases) => super.noSuchMethod(
         Invocation.method(
           #unbind,
           [useCases],
@@ -600,11 +637,313 @@ class MockProcessCameraProvider extends _i1.Mock
       );
 }
 
+/// A class which mocks [Recorder].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockRecorder extends _i1.Mock implements _i10.Recorder {
+  @override
+  _i14.Future<_i8.PendingRecording> prepareRecording(String? path) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #prepareRecording,
+          [path],
+        ),
+        returnValue:
+            _i14.Future<_i8.PendingRecording>.value(_FakePendingRecording_7(
+          this,
+          Invocation.method(
+            #prepareRecording,
+            [path],
+          ),
+        )),
+        returnValueForMissingStub:
+            _i14.Future<_i8.PendingRecording>.value(_FakePendingRecording_7(
+          this,
+          Invocation.method(
+            #prepareRecording,
+            [path],
+          ),
+        )),
+      ) as _i14.Future<_i8.PendingRecording>);
+}
+
+/// A class which mocks [PendingRecording].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockPendingRecording extends _i1.Mock implements _i8.PendingRecording {
+  @override
+  _i14.Future<_i9.Recording> start() => (super.noSuchMethod(
+        Invocation.method(
+          #start,
+          [],
+        ),
+        returnValue: _i14.Future<_i9.Recording>.value(_FakeRecording_8(
+          this,
+          Invocation.method(
+            #start,
+            [],
+          ),
+        )),
+        returnValueForMissingStub:
+            _i14.Future<_i9.Recording>.value(_FakeRecording_8(
+          this,
+          Invocation.method(
+            #start,
+            [],
+          ),
+        )),
+      ) as _i14.Future<_i9.Recording>);
+}
+
+/// A class which mocks [Recording].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockRecording extends _i1.Mock implements _i9.Recording {
+  @override
+  _i14.Future<void> close() => (super.noSuchMethod(
+        Invocation.method(
+          #close,
+          [],
+        ),
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
+  @override
+  _i14.Future<void> pause() => (super.noSuchMethod(
+        Invocation.method(
+          #pause,
+          [],
+        ),
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
+  @override
+  _i14.Future<void> resume() => (super.noSuchMethod(
+        Invocation.method(
+          #resume,
+          [],
+        ),
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
+  @override
+  _i14.Future<void> stop() => (super.noSuchMethod(
+        Invocation.method(
+          #stop,
+          [],
+        ),
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
+}
+
+/// A class which mocks [VideoCapture].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockVideoCapture extends _i1.Mock implements _i27.VideoCapture {
+  @override
+  _i14.Future<_i10.Recorder> getOutput() => (super.noSuchMethod(
+        Invocation.method(
+          #getOutput,
+          [],
+        ),
+        returnValue: _i14.Future<_i10.Recorder>.value(_FakeRecorder_9(
+          this,
+          Invocation.method(
+            #getOutput,
+            [],
+          ),
+        )),
+        returnValueForMissingStub:
+            _i14.Future<_i10.Recorder>.value(_FakeRecorder_9(
+          this,
+          Invocation.method(
+            #getOutput,
+            [],
+          ),
+        )),
+      ) as _i14.Future<_i10.Recorder>);
+}
+
+/// A class which mocks [BuildContext].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockBuildContext extends _i1.Mock implements _i11.BuildContext {
+  @override
+  _i11.Widget get widget => (super.noSuchMethod(
+        Invocation.getter(#widget),
+        returnValue: _FakeWidget_10(
+          this,
+          Invocation.getter(#widget),
+        ),
+        returnValueForMissingStub: _FakeWidget_10(
+          this,
+          Invocation.getter(#widget),
+        ),
+      ) as _i11.Widget);
+  @override
+  bool get mounted => (super.noSuchMethod(
+        Invocation.getter(#mounted),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
+  @override
+  bool get debugDoingBuild => (super.noSuchMethod(
+        Invocation.getter(#debugDoingBuild),
+        returnValue: false,
+        returnValueForMissingStub: false,
+      ) as bool);
+  @override
+  _i11.InheritedWidget dependOnInheritedElement(
+    _i11.InheritedElement? ancestor, {
+    Object? aspect,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #dependOnInheritedElement,
+          [ancestor],
+          {#aspect: aspect},
+        ),
+        returnValue: _FakeInheritedWidget_11(
+          this,
+          Invocation.method(
+            #dependOnInheritedElement,
+            [ancestor],
+            {#aspect: aspect},
+          ),
+        ),
+        returnValueForMissingStub: _FakeInheritedWidget_11(
+          this,
+          Invocation.method(
+            #dependOnInheritedElement,
+            [ancestor],
+            {#aspect: aspect},
+          ),
+        ),
+      ) as _i11.InheritedWidget);
+  @override
+  void visitAncestorElements(bool Function(_i11.Element)? visitor) =>
+      super.noSuchMethod(
+        Invocation.method(
+          #visitAncestorElements,
+          [visitor],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void visitChildElements(_i11.ElementVisitor? visitor) => super.noSuchMethod(
+        Invocation.method(
+          #visitChildElements,
+          [visitor],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void dispatchNotification(_i11.Notification? notification) =>
+      super.noSuchMethod(
+        Invocation.method(
+          #dispatchNotification,
+          [notification],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  _i13.DiagnosticsNode describeElement(
+    String? name, {
+    _i13.DiagnosticsTreeStyle? style = _i13.DiagnosticsTreeStyle.errorProperty,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #describeElement,
+          [name],
+          {#style: style},
+        ),
+        returnValue: _FakeDiagnosticsNode_12(
+          this,
+          Invocation.method(
+            #describeElement,
+            [name],
+            {#style: style},
+          ),
+        ),
+        returnValueForMissingStub: _FakeDiagnosticsNode_12(
+          this,
+          Invocation.method(
+            #describeElement,
+            [name],
+            {#style: style},
+          ),
+        ),
+      ) as _i13.DiagnosticsNode);
+  @override
+  _i13.DiagnosticsNode describeWidget(
+    String? name, {
+    _i13.DiagnosticsTreeStyle? style = _i13.DiagnosticsTreeStyle.errorProperty,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #describeWidget,
+          [name],
+          {#style: style},
+        ),
+        returnValue: _FakeDiagnosticsNode_12(
+          this,
+          Invocation.method(
+            #describeWidget,
+            [name],
+            {#style: style},
+          ),
+        ),
+        returnValueForMissingStub: _FakeDiagnosticsNode_12(
+          this,
+          Invocation.method(
+            #describeWidget,
+            [name],
+            {#style: style},
+          ),
+        ),
+      ) as _i13.DiagnosticsNode);
+  @override
+  List<_i13.DiagnosticsNode> describeMissingAncestor(
+          {required Type? expectedAncestorType}) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #describeMissingAncestor,
+          [],
+          {#expectedAncestorType: expectedAncestorType},
+        ),
+        returnValue: <_i13.DiagnosticsNode>[],
+        returnValueForMissingStub: <_i13.DiagnosticsNode>[],
+      ) as List<_i13.DiagnosticsNode>);
+  @override
+  _i13.DiagnosticsNode describeOwnershipChain(String? name) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #describeOwnershipChain,
+          [name],
+        ),
+        returnValue: _FakeDiagnosticsNode_12(
+          this,
+          Invocation.method(
+            #describeOwnershipChain,
+            [name],
+          ),
+        ),
+        returnValueForMissingStub: _FakeDiagnosticsNode_12(
+          this,
+          Invocation.method(
+            #describeOwnershipChain,
+            [name],
+          ),
+        ),
+      ) as _i13.DiagnosticsNode);
+}
+
 /// A class which mocks [TestInstanceManagerHostApi].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTestInstanceManagerHostApi extends _i1.Mock
-    implements _i24.TestInstanceManagerHostApi {
+    implements _i28.TestInstanceManagerHostApi {
   @override
   void clear() => super.noSuchMethod(
         Invocation.method(
@@ -615,10 +954,68 @@ class MockTestInstanceManagerHostApi extends _i1.Mock
       );
 }
 
+/// A class which mocks [TestSystemServicesHostApi].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockTestSystemServicesHostApi extends _i1.Mock
+    implements _i28.TestSystemServicesHostApi {
+  @override
+  _i14.Future<_i6.CameraPermissionsErrorData?> requestCameraPermissions(
+          bool? enableAudio) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #requestCameraPermissions,
+          [enableAudio],
+        ),
+        returnValue: _i14.Future<_i6.CameraPermissionsErrorData?>.value(),
+        returnValueForMissingStub:
+            _i14.Future<_i6.CameraPermissionsErrorData?>.value(),
+      ) as _i14.Future<_i6.CameraPermissionsErrorData?>);
+  @override
+  void startListeningForDeviceOrientationChange(
+    bool? isFrontFacing,
+    int? sensorOrientation,
+  ) =>
+      super.noSuchMethod(
+        Invocation.method(
+          #startListeningForDeviceOrientationChange,
+          [
+            isFrontFacing,
+            sensorOrientation,
+          ],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void stopListeningForDeviceOrientationChange() => super.noSuchMethod(
+        Invocation.method(
+          #stopListeningForDeviceOrientationChange,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  String getTempFilePath(
+    String? prefix,
+    String? suffix,
+  ) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getTempFilePath,
+          [
+            prefix,
+            suffix,
+          ],
+        ),
+        returnValue: '',
+        returnValueForMissingStub: '',
+      ) as String);
+}
+
 /// A class which mocks [ZoomState].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockZoomState extends _i1.Mock implements _i13.ZoomState {
+class MockZoomState extends _i1.Mock implements _i16.ZoomState {
   @override
   double get minZoomRatio => (super.noSuchMethod(
         Invocation.getter(#minZoomRatio),
@@ -633,202 +1030,62 @@ class MockZoomState extends _i1.Mock implements _i13.ZoomState {
       ) as double);
 }
 
-/// A class which mocks [BuildContext].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockBuildContext extends _i1.Mock implements _i8.BuildContext {
-  MockBuildContext() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i8.Widget get widget => (super.noSuchMethod(
-        Invocation.getter(#widget),
-        returnValue: _FakeWidget_7(
-          this,
-          Invocation.getter(#widget),
-        ),
-      ) as _i8.Widget);
-  @override
-  bool get mounted => (super.noSuchMethod(
-        Invocation.getter(#mounted),
-        returnValue: false,
-      ) as bool);
-  @override
-  bool get debugDoingBuild => (super.noSuchMethod(
-        Invocation.getter(#debugDoingBuild),
-        returnValue: false,
-      ) as bool);
-  @override
-  _i8.InheritedWidget dependOnInheritedElement(
-    _i8.InheritedElement? ancestor, {
-    Object? aspect,
-  }) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #dependOnInheritedElement,
-          [ancestor],
-          {#aspect: aspect},
-        ),
-        returnValue: _FakeInheritedWidget_8(
-          this,
-          Invocation.method(
-            #dependOnInheritedElement,
-            [ancestor],
-            {#aspect: aspect},
-          ),
-        ),
-      ) as _i8.InheritedWidget);
-  @override
-  void visitAncestorElements(bool Function(_i8.Element)? visitor) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #visitAncestorElements,
-          [visitor],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void visitChildElements(_i8.ElementVisitor? visitor) => super.noSuchMethod(
-        Invocation.method(
-          #visitChildElements,
-          [visitor],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  void dispatchNotification(_i8.Notification? notification) =>
-      super.noSuchMethod(
-        Invocation.method(
-          #dispatchNotification,
-          [notification],
-        ),
-        returnValueForMissingStub: null,
-      );
-  @override
-  _i10.DiagnosticsNode describeElement(
-    String? name, {
-    _i10.DiagnosticsTreeStyle? style = _i10.DiagnosticsTreeStyle.errorProperty,
-  }) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #describeElement,
-          [name],
-          {#style: style},
-        ),
-        returnValue: _FakeDiagnosticsNode_9(
-          this,
-          Invocation.method(
-            #describeElement,
-            [name],
-            {#style: style},
-          ),
-        ),
-      ) as _i10.DiagnosticsNode);
-  @override
-  _i10.DiagnosticsNode describeWidget(
-    String? name, {
-    _i10.DiagnosticsTreeStyle? style = _i10.DiagnosticsTreeStyle.errorProperty,
-  }) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #describeWidget,
-          [name],
-          {#style: style},
-        ),
-        returnValue: _FakeDiagnosticsNode_9(
-          this,
-          Invocation.method(
-            #describeWidget,
-            [name],
-            {#style: style},
-          ),
-        ),
-      ) as _i10.DiagnosticsNode);
-  @override
-  List<_i10.DiagnosticsNode> describeMissingAncestor(
-          {required Type? expectedAncestorType}) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #describeMissingAncestor,
-          [],
-          {#expectedAncestorType: expectedAncestorType},
-        ),
-        returnValue: <_i10.DiagnosticsNode>[],
-      ) as List<_i10.DiagnosticsNode>);
-  @override
-  _i10.DiagnosticsNode describeOwnershipChain(String? name) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #describeOwnershipChain,
-          [name],
-        ),
-        returnValue: _FakeDiagnosticsNode_9(
-          this,
-          Invocation.method(
-            #describeOwnershipChain,
-            [name],
-          ),
-        ),
-      ) as _i10.DiagnosticsNode);
-}
-
 /// A class which mocks [LiveData].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockLiveCameraState extends _i1.Mock
-    implements _i3.LiveData<_i12.CameraState> {
+    implements _i3.LiveData<_i15.CameraState> {
   MockLiveCameraState() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i11.Future<void> observe(_i25.Observer<_i12.CameraState>? observer) =>
+  _i14.Future<void> observe(_i29.Observer<_i15.CameraState>? observer) =>
       (super.noSuchMethod(
         Invocation.method(
           #observe,
           [observer],
         ),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
   @override
-  _i11.Future<void> removeObservers() => (super.noSuchMethod(
+  _i14.Future<void> removeObservers() => (super.noSuchMethod(
         Invocation.method(
           #removeObservers,
           [],
         ),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
 }
 
 /// A class which mocks [LiveData].
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockLiveZoomState extends _i1.Mock
-    implements _i3.LiveData<_i13.ZoomState> {
+    implements _i3.LiveData<_i16.ZoomState> {
   MockLiveZoomState() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i11.Future<void> observe(_i25.Observer<_i13.ZoomState>? observer) =>
+  _i14.Future<void> observe(_i29.Observer<_i16.ZoomState>? observer) =>
       (super.noSuchMethod(
         Invocation.method(
           #observe,
           [observer],
         ),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
   @override
-  _i11.Future<void> removeObservers() => (super.noSuchMethod(
+  _i14.Future<void> removeObservers() => (super.noSuchMethod(
         Invocation.method(
           #removeObservers,
           [],
         ),
-        returnValue: _i11.Future<void>.value(),
-        returnValueForMissingStub: _i11.Future<void>.value(),
-      ) as _i11.Future<void>);
+        returnValue: _i14.Future<void>.value(),
+        returnValueForMissingStub: _i14.Future<void>.value(),
+      ) as _i14.Future<void>);
 }