+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/legacy/web_kit_webview_widget.dart
@@ -5,7 +5,6 @@
 import 'dart:async';
 import 'dart:math';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:path/path.dart' as path;
@@ -316,7 +315,7 @@ class WebKitWebViewPlatformController extends WebViewPlatformController {
     final NSUrlRequest urlRequest = NSUrlRequest(
       url: request.uri.toString(),
       allHttpHeaderFields: request.headers,
-      httpMethod: describeEnum(request.method),
+      httpMethod: request.method.name,
       httpBody: request.body,
     );